Add files via upload
CALCULATOR WEB APPLICATION
This project is a simple yet functional calculator built using HTML, CSS, and JavaScript. It features a clean and intuitive user interface that allows for basic arithmetic operations.

Features:-
Basic Operations: Perform addition, subtraction, multiplication, and division.
Responsive Design: The layout is designed to be user-friendly on both desktop and mobile devices.
Clear and Delete: Easily clear the current input or delete the last entered digit.
Keyboard Support: Allows input via the keyboard for faster calculations.

Technologies Used:-
HTML: For structuring the web page.
CSS: For styling the calculator and making it visually appealing.
JavaScript: For implementing the calculator's logic and interactivity.
Usage :-
Simply open the index.html file in a web browser, and the calculator will be ready to use. Click on the buttons or use your keyboard to perform calculations.

Contributing :-
Feel free to fork this repository, make improvements, and submit pull requests. All contributions are welcome!